Add files via upload
For counting the legos, I have used the following methods

Segmentation by Colour
Bag of Visual Words - For Class Identification
Count the lego
Step 1: Segmentation by Color We are using the color space to segment the legos into blue and red separatrely After that we will be using the morphological operators to get information get then classified into individual objects and then we use Image labeling to label the segmented images The indidivual segmented images will be predicted to belong to blue or red legos based on the LegoCategoryClassifier built from the BagOfVIsual Words.

Step 2: Bag of Visual Words We are following the below steps to count the blue and red legos

Bag of Visual Words: The Bag of Visual words help you to classify object into into different categories. In this scenario, we are using it to categorise the BagofWords into a. 2x2 legos b. 2x4 legos c. Others
The main reason why we are clasifying it based on shape is because we are doing colour segmentation, so it it not necesarry to classify them based on colour again in BagOfWords. So we only need the shape information for classification.

Bag of Words Feature Extraction: For the feature extract of BagOf Visual Words Im using a custom extractor - HOG extractor to extract he features about the Images. You can see this in BagOfWords.m files which calls the HOG custom Extracor. The HOG Feature Extractor seems to work well for the Extraction since the objects are subject to lot of variation to illumination

We are using a custom template SVM where we give the parameters to tune the Model and obtain results from that. It is also part of BagOfWords.m file.

Then we are splitting the images into Training Set and Test Set and then testing the images to evaluate the score. This can also be seen when the BagOfWords.m file is executed separately. After evaluating we are save the CategoryClassifer as LegoCategoryClassifer.m file to be used in this function.

Step 3: Count the legos The legos will be counted based on the classifier and will sent as output.